Moved UINavigationButton extension to softUIView module
