+++ Wake On Lan/Modules/HostList/Sources/HostList/HostListView.swift
@@ -217,21 +217,6 @@ extension HostListView {
             symbol = .deviceSort
         }
         let image = UIImage(sfSymbol: symbol, withConfiguration: .init(weight: .semibold))
-        sortItemsButton.updateIconImageView(to: image)
-    }
-}
-
-extension UIBarButtonItem {
-    func updateIconImageView(to newImage: UIImage?) {
-        guard let customView = self.customView as? SoftUIView else {
-            return
-        }
-        let imageView = UIImageView(image: newImage)
-        imageView.tintColor = WOLResources.Asset.Colors.secondary.color
-        customView.configure(with: SoftUIViewModel(contentView: imageView))
-
-        imageView.snp.makeConstraints {
-            $0.edges.equalToSuperview().inset(6)
-        }
+        sortItemsButton.updateSoftUIImageView(to: image, tintColor: WOLResources.Asset.Colors.secondary.color)
     }
 }
